Move OAuthPrompt static methods and settings to Microsoft.Bot.Builder.OAuth.OAuthCardProvider, enabling use outside of dialogs
